+++ src/js/textMode/color/colorPaletteChoosePreset.js
@@ -68,8 +68,8 @@ ColorPaletteChoosePreset.prototype = {
     }
 
     if(this.uiComponent == null) {
-      var width = 636;
-      var height = 646;
+      var width = 800;
+      var height = 800;
       this.uiComponent = UI.create("UI.Dialog", { "id": "colorPaletteChoosePresetDialog", "title": "Choose A Colour Palette", "width": width, "height": height });
 
       this.splitPanel = UI.create("UI.SplitPanel");